Improved automation

+++ install.sh
@@ -2,6 +2,10 @@
 
 DIR=/media/mmcblk0p1/apps/sdr_transceiver_ft8
 
+mount -o rw,remount /media/mmcblk0p1
+
+$DIR/stop.sh
+
 chmod +x *.sh radioday upload-to-rbn
 
 cp temp.sh ~
@@ -13,11 +17,12 @@ cp radioday $DIR
 cp upload-to-rbn $DIR
 
 # Back up original decode-ft8.sh 
-test -e $DIR/decode-ft8.sh.orig || \
-	cp $DIR/decode-ft8.sh $DIR/decode-ft8.sh.orig
+test -e $DIR/decode-ft8.sh.orig || cp $DIR/decode-ft8.sh $DIR/decode-ft8.sh.orig
 cp decode-ft8.sh $DIR
 
 # Set FT8 receiver to run at boot
 cp $DIR/start.sh /media/mmcblk0p1
 
 lbu commit -d
+
+mount -o ro,remount /media/mmcblk0p1
